+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- new event `element:ready` triggered after an element and its children has been
+  fully constructed.
 - add plugin callbacks `init()` and `setup()`.
 - new rule `require-sri`.
 - add `<slot>` element.